src: add minimal code to load/convert/write a png file
png2gcode.c currently doesn't do much, it loads a PNG image (indexed,
gray or RGB, all turned to RGBA), converts it to grayscale, then to
work area, then back to gray and saves it to another file. It can also
support cropping a file to specific dimensions. No G-CODE is produced
for now.

This uses libpng's simplified API so it requires libpng >= 1.6.