src: add a left-right only raster mode
When engraving metal, it's common that the effect of the laser dot isn't
the same at the beginning and at the end of a line, either due to a stronger
pulse at the beginning, or to the heat accumulation while engraving. The
issue it causes in raster mode is that there is a small but noticeable
offset between left and right passes, that cannot easily be corrected.

This patch adds a lerf-right-only raster mode, which will be slower since
the head will have to fly over the artwork for every line, but will make
sure that dots are much better aligned. The mode is called "raster-lr".