src: add a diffusion ratio for soften and hash modes
When running in soften or in hash modes, we try to limit the impact of
adjacent pixels on each pixel. This new ratio set by -d indicates how
much of a pixel's energy diffuses into the next pixel around. It's set
to 0.18 by default, which makes it so that almost all of the pixel's
energy spreads over the 4 adjacent pixels and the 4 diagonal ones.

Combined with -H, this further improves rastering accuracy because
instead of leaving white spots in on every other pixel, it can still
send a bit of energy there to compensate for what is still missing
from the neighbors.

A more elaborate method would consist in using 4 passes:

  1) compute the energy sent on each 1/2 point, which spreads in
     prev and next pixels (both stored separately for each of them)

  2) compute a correction dose for each of the empty pixel based on this:
      if pre > target: prev.excess += diff
      if post > target: next.excess += diff
      if pre+post > target: val=target-pre-post; prev.excess += val*pre/(pre+post); next.excess += val*post/(pre+post);

  3) adjust neighbors power in order to balance the error on their neighbours

  4) complete the empty points based on the missing excess, taking into
     account that the pre-dose will darken the spot and make the wood
     absorb light faster.

Note that the amount of black varies as the black surface in the hole,
which varies with the square of the radius while the amount of energy varies
with the cube of the radius (since all the matter has to be vaporized), thus
the black intensity varies with PWM^(2/3) (or PWM varies with black^(3/2)).