src: add a new histogram-based quantization (-Q)
With -Q, the quantization will be performed based on the pixel
luminosity's frequency. The goal is to try to better auto-contrast
images. Maybe we could get rid of the quantization on output and
leave it to -q after only emitting a value related to the frequency,
that's still to be checked. In this case we could also automatically
rescale values according to min/avg/max, and even use both at choice
arithmetic or geometric mean. At least for now it already gives good
results.