src: add a new "twins" processing
In this mode, pixels are considered in pairs, and we'll try to
maximize the difference between their values while still offering the
full range. If their average value is between 0..0.5, one pixel is clear
and the other one is proportionally filled; between 0.5..1.0 one pixel is
filled and the other one is proportionally filled. This effectively makes
adjacent pixels twice as large on average but with a fabric aspect that
covers for the lack of linearity of anodized aluminum. It seems to work
well with gamma value 0.25.