src: add a new centering technique (circle)
This adds a new centering technique consisting in finding the center of
the smallest circle that contains the whole image. The principle is first
to eliminate all non-boundary pixels, then to proceed like a mechanical
circular piece would act if tightened around a material: apply pressure
on the longest side and move the center the other direction, until there's
no more move possible because the smallest circle was reached. This can be
simplified by starting from the center and going into the direction that
shows the smallest max_radius to the center.

The max computed radius is reported, so that it's easier to figure the
most optimal scale. This will significantly ease printing irregular
images on circular pieces.