src: fix problem of output precision and bad format
When some values get too close to zero, the "%g" output will switch to
the "%e" format. This typically happens with a centered image when y
can get extremely close to zero. A value such as "8e-5" may be emitted,
causing a parsing error and/or a jump to the wrong place.

Let's introduce a float rounding function for 4 digits, called "f4",
which we pass all printf arguments through. This cleans up the output
and we can now read:

   G0 X-15.2262 Y0.0001

instead of:

   G0 X-15.2262 Y8.201599e-05