src: add a new --rl-shift option to set the machine's L->R shift
Probably due to some parts precision or elasticity, there's always a small
shift between L->R passes and R->L ones. This option permits to apply a
correction to this lack of precision which is a problem when printing
hashed areas. It defaults to -0.05mm which is what I need on my EleksMaker
A3 Pro.