src: add a test mode which generates a hashed area
The principle is to help calibrate the laser focus by generating a hashed
area directly here without having to use complicated painting tools. The
size is passed in argument and it fills the work area just as if a file
had been loaded. We may imagine writing text there or adding gradients
later.