src: support special handling of the beam's width
There's always some over-burning of contrasted areas because the beam is
very wide compared to the pixel's width, so the left and write margins
are exceeded by the beam's radius. Thus if the beam's width is equal to
the pixel's width, it's impossible to alternate white and black pixels,
they'll be completely burnt. This is what the hashed mode aims at
addressing, but is not sufficient to emit small pixels.

What we do here is to define the beam's width. It's always limited to
90% the pixel's width anyway. As such, when set, a large dark area will
consist in short pulses that only try to cover the difference between
the pixel's width and the beam's width.

In order to partially eliminate the risk of vertical stripes, in LR
mode we apply the beam's width as an offset every other line. The beam
width remains zero by default.