bug: fix the gamma calculation
The gamma calculation that aimed to be approximately centered on the X
axis always had an Y offset at the bottom or at the top equal to the
absolute value of the gamma value, resulting in excessive burns for very
low values. Let's get back to the standard algorithm first.