src: support calculating gamma from the bright colors instead
Gamma processing is a compression at the bottom of the curve. But in
some cases it could be interesting to apply the compression at the top,
e.g. if printing reversed, or with certain materials. Here we support
gamma correction by flipping the X axis so that the correction compresses
the brightest colors.