src: add an acceleration margin on the X axis
some raster outputs are imperfect on the left and right edges due to the
acceleration and deceleration of the head. With M3 it causes overburns and
with M4 it causes underburns as the firmware tries hard to compensate. The
better solution consists in making it possible to enter the machine's
acceleration distance with -A on the command line so that the head starts
before this distance and finishes after. This guarantees that the speed
will not change in these areas. For large jumps in the middle, we also
apply the same algorithm and we don't perform them if the distance is not
at least 2*this margin (it was 20 pixels previously).