src: fix incorrect back-and-forth jumps when using beam width and acceleration
The beam width causes the termination of a pixel once drawn, and the return
to idle state, but that didn't cope well with acceleration that was forcing
a jump back to origin before displaying the new pixel. The cause was that
the x0 value was not reset. Resetting it to the last printed pixel seems to
do the job. It's worth noting that the entire algorithm would deserve being
redone to follow a Z-shape path looking only for non-zero areas and skipping
the zero ones as fast as possible (i.e. S0 or G0 depending on length).