src: add new transformation method "vertical"
It's similar to twins except that we do not alternate pixels on even/odd lines.
This results in burning along vertical stripes and leaving some material
unaffected, which shows much higher contrasts on weak materials like cardboard
by not losing the small pieces of burnt material that normally carry the color.
This requires a perfect alignment (raster-lr or optimal rl-delay).

This was found by accident by using an inappropriate rl-delay value with twins
that resulted in odd/even lines to be perfectly aligned and in the contrast to
be optimal. While -t leaves light gray, -V gives dense black. It doesn't give
good results with pixel smaller than the dot though, unless raster-lr is used
with lower power.