src: add a max-length option to limit long runs that cause burns
When engraving FS-dithered images, it's possible to see long burnt lines,
especially when engraving good which easily catches on fire. The new
--max-length option will enforce periodic disabled pixels. They will
try to be placed at more or less random locations to avoid vertical lines,
and lines will be alternated left and right when placing these dots in
order to further decrease the risk of too aligned lines.