src: parse new text-related parameters to create a new xfrm
This adds the ability to create a new XFRM_TEXT that consists in adding
text to the image. The position may be specified using b,c,t for bottom,
center, top, and l,c,r for left, center, right. New options might appear
in the future to automatically extend the frame for writing outside.

The zoom factor, FG alpha and value, BG alpha and value can also be
specified for any future text and are recorded in the xfrm.

If multiple lines of text are passed without new position specs nor
characteristics, they're assumed to be a same multi-line block and are
appended to the previous block.