src: add a 5x7 font definition and a test
This is a 5x7 font that was reimplemented from the HD44780 manual for
charset A02. A few characters were slightly modified to improve their
readability ('3', '7', '@', 'S', 'g', 'i', 'q', '~'). Overall this charset
was already very good for small pixel-based outputs so it made sense to
restart from this one. Only the 96 chars from 32 to 127 were implemented.

The array was defined as a static one in a .h file so that it's easy to
use from multiple programs if needed. A simple test program will print
each char from a string passed in argument, or print all of them 8 per
line if no argument is given, using '#' for each pixel. It renders quite
well in 'tiny' xterm font.