src: implement the text transformation
This implements the text as a transformation so that it can be passed
anywhere in the chain (before/after correction etc). Characters other
than 32..127 are replaced with a space. Special care is taken to use
the background color for one color after each char but only if it's not
the last char, and one line after the last text line but only if it's
not the last line.