src: fix character escaping again
Insufficient characters were escaped. A number of shell delimiters were
not (such as brackets, braces, pipes etc). Now we split them in two
categories, the unprintable ones that require $'\x' and the printable
ones that require quotes. It was verified that 0x20-0x7f are now correctly
printed and that the output gives the same values then fed back.

It's still not pretty because quoting is made a per-character while it
could easily be done as a block.

Maybe an improvement would be to first call a function to determine if
any char in a word requires escaping or quoting, and if so, to quote or
escape the whole string.