src: fix beam width processing regarding levels
We want the beam width action to make the stronger parts smaller compared
to weaker parts, not just 1 vs 0. Also the right-to-left walk was incorrect
and was using the wrong offset. Now it's possible to shrink the beam width
to very narrow values, these might even be usable for gradients in the
future.