src: support a uniform beam width mode
With --raster-ubw, the user can choose the width during which the beam
is turned on for each pixel. It means that long series of contiguous
pixels will be broken into individual dots followed by silences. This
can permit to avoid overburn when the beam is too wide and to create a
similar effect to what the hash tries to solve, e.g. on wood or paint.