src: simplify the raster output production using a queue
The logic used to emit the pixels is quite complicated because we always
have to know the state of the currently ongoing pixel. Let's implement a
move queue to merge adjacent moves with the same spindle.

The queue contains a source and a destination for the move, with the
spindle speed and G mode. There are still limitations, though, as we
don't know if the previous move used the same G mode, so we should
instead store the G_mode into the from/to vectors as well, to remind
which one was used last. In this case it could even make sense to
record the last spindle, though it's rare to emit multiple explicit
spindle values.

It's worth noting that this removes the need for flushing pending moves
from the next Y position, and also voids the need for the copy of the
last spindle value in curr_spindle.