src: add support for a "zero" transformation to zero the lowest value
When processing certain images with -a and -m, it often happens that the
output is not exactly zero, or that adding an offset to zero the output
will affect the output for higher values.

This new transformation simply finds the lowest value in the image and
assigns it the value zero, considering that it's not expected to be
etched. This can significantly speed up the engraving of areas with
large margins (e.g. a circle) as it will permit the beam to travel at
full speed on the area not supposed to be engraved.