src: store the G mode into the from/to locations of the queue
This is in order to know what the previous G mode was and avoid emitting
it with each and every move. It also makes sure that we detect that we
were still in init state, so that we now properly emit both X and Y for
the first G0 move.