src: switch positions to signed ints
x/y positions were stored as unsigned to accomodate the image dimensions
but in the end it's complicated as it doesn't permit them to be negative
(e.g. when traveling over the acceleration area). And changing just a
few inevitably yields plenty of signedness warnings. Let's just turn all
of them to signed at once, and add the missing range checks everywhere
required.