src: add generation date in the output file
When applying updates to arguments it's difficult to know if the file
opened and possibly refreshed in gcode sending utility is the up-to-date
one or not. Just adding the date on the first line is sufficient to see
it change in the console. It may also help figure under which condition
or for which project a gcode file was produced.