src: do not flush the queue for 0-length moves
Sometimes due to the margin calculations or accelerations, we can
inconditionally emit some 0-length moves where only the spindle
changes (e.g. send X12 S0 immediately after X12 S255). It makes
no sense to do that, as it artificially inflates the output file
and the work for the sender. Let's simply drop them without
updating the queue so that subsequent moves have a chance to be
merged.